refactor(chat-dataflow): retire chat_consumer + pipeline.stream_chat, add chat_node observability
Phase 5a Task 12 — final cleanup of the legacy chat path:

- new app/chat/agent_stream.py extracted from app/chat/pipeline.py:
  _build_and_stream + _BotCtx + _load_bot_context + _MAIN_CFG. Deviation
  from plan (approved): a separate module instead of inlining into
  chat_node.py to keep both files within CLAUDE.md's 300-line guideline.
  pre-safety race + buffer/queue logic (only ever used by stream_chat)
  is dropped — chat_node uses _resolve_pre_safety_for_part at segment
  boundaries.

- delete app/workers/chat_consumer.py and app/chat/pipeline.py.
  app/main.py no longer launches start_chat_consumer; chat_request
  queue is now driven by the runtime durable consumer wired to chat_node.
  chat/__init__.py drops the stream_chat export.

- chat_node now ports the ops observability lost when Task 10 plan-ported
  the main loop:
    * t_first_token tracked, CHAT_FIRST_TOKEN.observe on stream end
    * CHAT_PIPELINE_DURATION{stage=total} on happy-path final emit
    * structured chat_request_done log
  Metrics fire on the happy path only — pre-safety BLOCK and fetch-empty
  branches skip them (no tokens to measure). queue_wait_ms / publish_ms
  from the MQ consumer are dropped (no separate publish phase in graph).

- relocate StreamState / handle_token tests from the deleted
  test_pipeline.py to a new tests/unit/chat/test_stream.py (those test
  app/chat/stream.py, which still exists). _buffer_until_pre race tests
  removed alongside the helper. tests/unit/workers/test_chat_consumer.py
  removed; route_chat_node tests already cover redelivery + fan-out +
  proactive paths.
